refactor : move DateTimeExtensions to utils package and rename the MockChatRepository.kt to FakeChatRepository.kt
