+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/chat/utils/DateTimeExtensions.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.core_chat.data.chat
+package net.thechance.mena.core_chat.data.chat.utils
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
